Add more user-friendly output for down
Previously we always down'ed a service regardless of whether or not it
is enabled.  This is because sometimes services get disabled while
running, and we still want them to come down.

This patch makes it so we still put them down, but now instead of
printing success (which can be confusing), we print a disabled message.
We still check to make sure the service is down, and display an error if
it isn't