Bump version to 0.3.0
This introduces support for Mac OS.  However this is not well tested
because I don't have a Mac.  Will not be advertising the OS X support
until it's been tested by somebody.

Also want to look for regressions on Linux with the resolv services and
dns.  This version will not be pushed to ruby gems until some additional
testing has been done.  Thus, the v0.3.0 tag may not point to this
commit.