Fix broken custom ssl certs
If we are mounting ssl certs, use the codekitchen image directly.  This
comes with the downside that regular http will be broken, but until we
have time to fix the proxy image itself, this is at least better than
nothing.

If we are not mounting ssl certs, continue to use the dory image so that
http and https both work seamlessly.

Bump version to 0.3.11

Fixes #5