Handle interference from systemd services
On Ubuntu (and possibly any other distros that use systemd and/or
NetworkManager) there is a race condition that prevents dory from
starting the dnsmasq container.  This commit adds the capability of
putting those services down before starting the dnsmasq container.  This
will eliminate the race condition and allow dory to start properly.

However we don't want to just leave the services down, since that may
kill all networking.  So we will put the services back up when done
initializing.

This also adds a setting to allow you to skip the prompts, since the
prompts will likely get annoying.  You can also disable this feature
altogether in the same setting (will add something to the README about
this)